enhance(frontend): 設定の一括リセット機能をその他に移動
